+++ warp/native/tile_reduce.h
@@ -401,10 +401,7 @@ template <int Axis, typename Op, typename Tile> CUDA_CALLABLE_DEVICE auto tile_r
     WP_PRAGMA_UNROLL
     for (int i = 0; i < OutputRegLayout::NumRegs; ++i) {
         int linear = OutputRegLayout::linear_from_register(i);
-        if (!OutputRegLayout::valid(linear))
-            break;
-
-        output.data[i] = output_buffer[linear];
+        output.data[i] = OutputRegLayout::valid(linear) ? output_buffer[linear] : T {};
     }
 
     return output;