+++ warp/tests/unittest_suites.py
@@ -222,6 +222,7 @@ def default_suite(test_loader: unittest.TestLoader = unittest.defaultTestLoader)
     from warp.tests.tile.test_tile_atomic_bitwise import TestTileAtomicBitwise
     from warp.tests.tile.test_tile_cholesky import TestTileCholesky
     from warp.tests.tile.test_tile_cholesky_no_mathdx import TestTileCholeskyNoMathDx
+    from warp.tests.tile.test_tile_empty import TestTileEmpty
     from warp.tests.tile.test_tile_fft import TestTileFFT
     from warp.tests.tile.test_tile_fft_no_mathdx import TestTileFFTNoMathDx
     from warp.tests.tile.test_tile_func_arg import TestTileFuncArg
@@ -365,6 +366,7 @@ def default_suite(test_loader: unittest.TestLoader = unittest.defaultTestLoader)
         TestTileAtomicBitwise,
         TestTileCholesky,
         TestTileCholeskyNoMathDx,
+        TestTileEmpty,
         TestTileFFT,
         TestTileFFTNoMathDx,
         TestTileFuncArg,